requirements.txt
Babel==2.6.0
Django==1.11
PyYAML==3.13
boto3==1.9.0
botocore==1.12.1
dj-database-url==0.5.0
django-babel==0.6.2
django-bootstrap4==0.0.7
django-chartjs==1.3
django-crispy-forms==1.7.2
django-forms-bootstrap==3.1.0
django-formtools==2.1
django-model-utils==3.1.2
django-storages==1.7.1
djangorestframework==3.8.2
docutils==0.14
gunicorn==19.9.0
jmespath==0.9.3
prices==1.0.0
psycopg2==2.7.5
py-moneyed==0.7.0
python-dateutil==2.7.3
pytz==2018.5
s3transfer==0.1.13
stripe==2.7.0
typing==3.6.6
urllib3==1.23
